requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
django-parler>=2.3,<3.0
